chore: remove obsolete replication revision-one optimization
With #993, we introduced an optimization to retrieve documents via
`_all_docs` if they had a revision of 1, since there was no `_bulk_get`
API at that time. This optimization allowed us to retrieve many
revision-one documents with a single request, instead of requesting each
document individually. Documents with higher revisions still had to be
fetched one by one.

Since then, `_bulk_get` has been implemented in CouchDB and is used by
PouchDB's replicator, so this optimization is no longer necessary. On
the contrary, it now leads to additional requests that slow down
replication, although not that much.